mesa: *HACK* include sha1 implementation in mesa-sha1.c
This is done so that I can forget about sha1 for now and
concentrate on the other parts. It seems problematic to link against
any sha1 library in android. One in bionic seems not exposed (?) and
one in boringssl does not have stable API and is anyway different API
than one written against in Mesa .. I'm not sure what should be done.

Signed-off-by: Tapani Pälli <tapani.palli@intel.com>